fix: keep supervisor listen socket reusable across restarts
Bind with SO_REUSEPORT and ship an overlap-restart helper so launchd
KeepAlive no longer equals a multi-minute outage (#125).
